Refactor timeouts, make Test an abstract class
