+++ src/tests/ceramic_cas.ts
@@ -1,12 +1,14 @@
-import {assert, Services} from "../index";
+import {assert, Services, Test} from "../index";
 import { TileDoctype } from "@ceramicnetwork/doctype-tile"
 
 async function delay(mills: number): Promise<void> {
     await new Promise(resolve => setTimeout(() => resolve(), mills))
 }
 
-export class CeramicCasTest {
-    public async runTest(services: Services): Promise<void> {
+const TEST_TIMEOUT_MS = 1000 * 5 // 5 seconds
+
+export class CeramicCasTest extends Test {
+    async _runTest(services: Services): Promise<void> {
         const {ceramic} = services
 
         const initialContent = { foo: 'bar' }
@@ -18,5 +20,9 @@ export class CeramicCasTest {
         while(true) {
             await delay(1000)
         }
+    }
+
+    _getTestTimeout(): number {
+        return TEST_TIMEOUT_MS
     }
 }
